% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHD_navigate.R
\name{addType}
\alias{addType}
\title{Adds the type attribute for co-located POIs of multiple themes}
\usage{
addType(new_POIs, POIs, IDfield, bind)
}
\arguments{
\item{new_POIs}{(sf data.frame) new POIs to be tested against existing}

\item{POIs}{(sf data.frame) Existing POIs}

\item{IDfield}{(character) Name of 'Type' field to be modified in existing POI}

\item{bind}{(logical) whether to bind non co-located POIs to data frame or just
attribute existing POIs}
}
\value{
(sf data.frame) xisting POIs with Type field modified for
}
\description{
Adds the type attribute for co-located POIs of multiple themes
}
