% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_flow_line}
\alias{build_flow_line}
\title{Build LINESTRING from 2 geoms}
\usage{
build_flow_line(geom1, geom2)
}
\arguments{
\item{geom1}{(MULTILINESTRING) object 1}

\item{geom2}{(MULTILINESTRING) object 2}
}
\value{
a \code{sf} object
}
\description{
Build LINESTRING from 2 geoms
}
