% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_node_net}
\alias{build_node_net}
\title{Build Node Net}
\usage{
build_node_net(flowpaths, add_type = TRUE)
}
\arguments{
\item{flowpaths}{a filepath to a \code{parquet} file or \code{sf} object}

\item{add_type}{}
}
\value{
list
}
\description{
Build Node Net
}
