% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_prep_RPU.R
\name{build_rpu_parquet}
\alias{build_rpu_parquet}
\title{Build RPU Parquet file
Relies on have NHDPlus RDS files local}
\usage{
build_rpu_parquet(
  rpu = "01a",
  rds_dir = "/Users/mjohnson/nhd_rds",
  outdir = "./data",
  overwrite = FALSE
)
}
\arguments{
\item{rpu}{RPU to extract}

\item{rds_dir}{RDS files}

\item{outdir}{Location to save data}

\item{overwrite}{Should exisitng files be overwritten?}
}
\value{
list to files
}
\description{
Build RPU Parquet file
Relies on have NHDPlus RDS files local
}
