% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelpath_fill_in.R
\name{consolidate_levelpaths}
\alias{consolidate_levelpaths}
\title{Consolidate LevelPaths}
\usage{
consolidate_levelpaths(network, runner)
}
\arguments{
\item{network}{a network list}

\item{runner}{set of levelpaths to fill and correct}
}
\description{
Consolidate LevelPaths
}
