% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp_levelpath_fill2.R
\name{count_fps_ls}
\alias{count_fps_ls}
\title{Count flowpath linestrings}
\usage{
count_fps_ls(fp)
}
\arguments{
\item{fp}{\code{sf} object}
}
\value{
\code{sf} object
}
\description{
Count flowpath linestrings
}
