% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release-strategy.R
\name{create_release}
\alias{create_release}
\title{Create Release}
\usage{
create_release(fl, cat, dir, version)
}
\arguments{
\item{fl}{flowlines sf object}

\item{cat}{catchment sf object}

\item{dir}{output directory}

\item{version}{named release versions}
}
\description{
Create Release
}
