% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHD_navigate.R
\name{DS_poiFix}
\alias{DS_poiFix}
\title{Moves POI Upstream or downstream if it falls on COMID
@param POIs_wgeom (sf data.frame) POIs
@param nhdDF  (sf data.frame) valid data frame of NHD flowlines}
\usage{
DS_poiFix(POIs_wgeom, nhd)
}
\value{
(sf data.frame) data.frame of POIs with new COMID associated
}
\description{
Moves POI Upstream or downstream if it falls on COMID
@param POIs_wgeom (sf data.frame) POIs
@param nhdDF  (sf data.frame) valid data frame of NHD flowlines
}
