% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelpath_fill_in.R
\name{fill_level_path}
\alias{fill_level_path}
\title{Fill disconnected LevelPaths}
\usage{
fill_level_path(lpID, network)
}
\arguments{
\item{lpID}{levelpath ID to fill}

\item{network}{network}
}
\description{
Fill disconnected LevelPaths
}
