% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp_levelpath_fill2.R
\name{fill_level_path2}
\alias{fill_level_path2}
\title{Fill disconnected LevelPaths}
\usage{
fill_level_path2(lpID, network)
}
\arguments{
\item{lpID}{levelpath ID to fill}

\item{network}{network}
}
\value{

}
\description{
Fill disconnected LevelPaths
}
