% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelpath_fill_in.R
\name{find_disconnected_lp}
\alias{find_disconnected_lp}
\title{Find LevelPaths with holes}
\usage{
find_disconnected_lp(flowpaths)
}
\arguments{
\item{flowpaths}{a network obj}
}
\value{
vector of dicconected levelpaths
}
\description{
Find LevelPaths with holes
}
