% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_node}
\alias{find_node}
\title{Find flowline node
Returns the start or end node from a flowline.}
\usage{
find_node(x, position = "end")
}
\arguments{
\item{x}{\code{sf} flowline object}

\item{position}{Node to find: "start" or "end" or "both"}
}
\description{
Find flowline node
Returns the start or end node from a flowline.
}
