% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_orphan_flowpaths.R
\name{find_orphan_flowpaths}
\alias{find_orphan_flowpaths}
\title{Find Orphan Flowlines}
\usage{
find_orphan_flowpaths(files)
}
\arguments{
\item{files}{a list of filepaths (see \code{build_rpu_parquet})}
}
\description{
Find Orphan Flowlines
}
