% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_utils.R
\name{flowpaths_to_linestrings}
\alias{flowpaths_to_linestrings}
\title{Convert MULITLINESTINGS to LINESTRINGS}
\usage{
flowpaths_to_linestrings(flowpaths)
}
\arguments{
\item{flowpaths}{a flowpath \code{sf} object}
}
\value{
a \code{sf} object
}
\description{
Convert MULITLINESTINGS to LINESTRINGS
}
