% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_orphan_flowpaths.R
\name{merge_orphan_flowpaths}
\alias{merge_orphan_flowpaths}
\title{Merge Orphan Flowlines}
\usage{
merge_orphan_flowpaths(orphans)
}
\arguments{
\item{orphans}{a list of orphan flowpaths (see \code{find_orphan_flowpaths})}
}
\description{
Merge Orphan Flowlines
}
