% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{needs_layer}
\alias{needs_layer}
\title{needs_layer}
\usage{
needs_layer(db, layer)
}
\arguments{
\item{db}{character geopackage to check}

\item{layer}{character layer name}
}
\value{
logical
}
\description{
Checks if layer existing in geopackage
}
