% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHD_navigate.R
\name{NetworkConnection}
\alias{NetworkConnection}
\title{Identifies and connects dangles in network generated by Network Nav function}
\usage{
NetworkConnection(incom, nhd)
}
\arguments{
\item{inCOM}{(list) list of input COMIDs}

\item{nhdDF}{(sf data.frame) (data frame) valid data frame of NHD flowlines}

\item{withTrib}{(logical) flag for if the upstream navigation should include tributaries}
}
\value{
(list) list of COMIDs connecting dangle to existing network
}
\description{
Identifies and connects dangles in network generated by Network Nav function
}
