% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHD_navigate.R
\name{NetworkNav}
\alias{NetworkNav}
\title{Network navigation for upstream/downstream from a COMID of interest}
\usage{
NetworkNav(inCom, nhdDF, withTrib)
}
\arguments{
\item{nhdDF}{(sf data.frame) (data frame) valid data frame of NHD flowlines}

\item{withTrib}{(logical) flag for if the upstream navigation should include tributaries}

\item{inCOM}{(list) list of input COMIDs}
}
\value{
(list) list of COMIDs upstream of point
}
\description{
Network navigation for upstream/downstream from a COMID of interest
}
