% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHD_navigate.R
\name{POI_creation}
\alias{POI_creation}
\title{Creates POIs for a given data theme}
\usage{
POI_creation(srcData, nhdDF, IDfield)
}
\arguments{
\item{srcData}{(data.frame) (data frame) DF with two columns:}

\item{nhdDF}{(sf data.frame) valid data frame of NHD flowlines}

\item{IDfield}{character) Name of 'Type' field to be modified in POI}
}
\value{
(sf data.frame) OIs for the specific data theme
}
\description{
Creates POIs for a given data theme
}
