% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHD_navigate.R
\name{POI_move_down}
\alias{POI_move_down}
\title{Collaposes/merges POIs together based on drainage area ratio and data theme
@param out_gpkg (geopackage) output geopackage to write intermediate results to
@param POIs  (sf data.frame) Original  POIs
@param Seg  (sf data.frame) dissolved segments
@param Seg2  (sf data.frame) (data frame) raw segments/flowlines
@param exp  (string) Type of thematic POI being moved round
@param DA_diff  (float) threshold of drainage area difference to}
\usage{
POI_move_down(out_gpkg, POIs, Seg, Seg2, exp, DA_diff)
}
\value{
(list of data.frames) 1 - New set of POIs
}
\description{
Collaposes/merges POIs together based on drainage area ratio and data theme
@param out_gpkg (geopackage) output geopackage to write intermediate results to
@param POIs  (sf data.frame) Original  POIs
@param Seg  (sf data.frame) dissolved segments
@param Seg2  (sf data.frame) (data frame) raw segments/flowlines
@param exp  (string) Type of thematic POI being moved round
@param DA_diff  (float) threshold of drainage area difference to
}
