% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_network_graph}
\alias{prep_network_graph}
\title{Build HydroNetwork Graph}
\usage{
prep_network_graph(flowpaths, catchments, ID_col = "ID", directed = FALSE)
}
\arguments{
\item{flowpaths}{a filepath to a \code{parquet} file or \code{sf} object}

\item{catchments}{a filepath to a \code{parquet} file or \code{sf} object}

\item{ID_col}{the desired ID column}

\item{directed}{should graph be directed?}
}
\description{
Build HydroNetwork Graph
}
