% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHD_navigate.R
\name{segment_creation}
\alias{segment_creation}
\title{Creates finalized segments and routing
@param nhdDF (sf data.frame) valid data frame of NHD flowlines
@param routing_fix  (sf data.frame) any additional routing fixes}
\usage{
segment_creation(nhdDF, routing_fix)
}
\value{
(sf data.frame) data.frame of segments
}
\description{
Creates finalized segments and routing
@param nhdDF (sf data.frame) valid data frame of NHD flowlines
@param routing_fix  (sf data.frame) any additional routing fixes
}
