% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHD_navigate.R
\name{segment_increment}
\alias{segment_increment}
\title{Creates raw and dissolved segment layers with necessaary
@param nhdDF (sf data.frame) valid data frame of NHD flowlines
@param POIs  (sf data.frame) Existing POIs}
\usage{
segment_increment(nhdDF, POIs)
}
\value{
(sf data.frame) data.frame of segments connecting POIs attributed
with POI_ID for each upstream flowpath
}
\description{
Creates raw and dissolved segment layers with necessaary
@param nhdDF (sf data.frame) valid data frame of NHD flowlines
@param POIs  (sf data.frame) Existing POIs
}
