% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelpath_fill_in.R
\name{spLineMerge}
\alias{spLineMerge}
\title{rgeos LINE Merge}
\usage{
spLineMerge(lines, ID)
}
\arguments{
\item{lines}{lines to merge}

\item{ID}{ID to merge over}
}
\value{
'sf object
}
\description{
Wayyyy faster then either data.table, or sf based line merging
}
