% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHD_navigate.R
\name{structPOIsNet}
\alias{structPOIsNet}
\title{Identifies and attributes structural POIs
@param nhdDF (sf data.frame) valid data frame of NHD flowlines
@param final_POIs  (sf data.frame) final POIs}
\usage{
structPOIsNet(nhdDF, final_POIs)
}
\value{
(data.frame columns) 1/0 Columns indicating structural POIs and the structural network
}
\description{
Identifies and attributes structural POIs
@param nhdDF (sf data.frame) valid data frame of NHD flowlines
@param final_POIs  (sf data.frame) final POIs
}
