% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHD_navigate.R
\name{switchDiv}
\alias{switchDiv}
\title{Switches valid POIs from minor to major path divergences}
\usage{
switchDiv(insegs, nhdDF)
}
\arguments{
\item{nhdDF}{(sf data.frame) (data frame) valid data frame of NHD flowlines}

\item{inSegs}{(list) list of input COMIDs representing POIs}
}
\value{
(sf data.frame) Corresponding major path COMID for POI
}
\description{
Switches valid POIs from minor to major path divergences
}
