% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHD_navigate.R
\name{VPU_Subset}
\alias{VPU_Subset}
\title{Subsets VPU for terminal outlets}
\usage{
VPU_Subset(nhd_rds, VPU, regoutlets = NULL)
}
\arguments{
\item{nhd_rds}{path to staged nhd flowline rds file}

\item{VPU}{VPU from NHDPlusV2}

\item{new_atts}{updated nhdplus attributes}

\item{nhdpath}{path to nhdplus geodatabase}
}
\value{
sf data.frame containing subset VPU
}
\description{
Subsets VPU for terminal outlets
}
\author{
David Blodgett
}
