% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_network_gpkg}
\alias{write_network_gpkg}
\title{Write a HydroNetwork GPKG}
\usage{
write_network_gpkg(network, outpath)
}
\arguments{
\item{network}{a newtwork list containing a cat, fl, and nex item (see \code{prep_network_graph})}

\item{outpath}{where to write the file?}
}
\value{

}
\description{
Write a HydroNetwork GPKG
}
