% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelpath_fill_in.R
\name{st_anti_filter}
\alias{st_anti_filter}
\title{Anti filter}
\usage{
st_anti_filter(.x, .y, .predicate = st_intersects)
}
\arguments{
\item{.x}{\code{sf} object}

\item{.y}{\code{sf} object}

\item{.predicate}{spatial predicate}
}
\value{
\code{sf} object
}
\description{
Anti filter
}
